transformer/requirements.txt

torch>=1.13.0
torchvision>=0.14.0
matplotlib>=3.5.0
numpy>=1.21.0
requests>=2.28.0
tqdm>=4.64.0
pyyaml>=6.0
